+++ demo/table_recognition/lgpma/tools/test_pub.py
@@ -1,0 +1,52 @@
+"""
+##################################################################################################
+# Copyright Info :    Copyright (c) Davar Lab @ Hikvision Research Institute. All rights reserved.
+# Filename       :    test_pub.py
+# Abstract       :    Script for inference
+
+# Current Version:    1.0.0
+# Date           :    2021-09-18
+##################################################################################################
+"""
+
+import json
+import jsonlines
+import numpy as np
+from tqdm import tqdm
+from eval_pub.metric import TEDS
+from eval_pub.format import format_html
+from davarocr.davar_common.apis import inference_model, init_model
+
+# path setting
+savepath = "/path/to/save/prediction" # path to save prediction
+config_file = '/path/to/config' # config path
+checkpoint_file = '/path/to/model/' # model path
+
+# loading model from config file and pth file
+model = init_model(config_file, checkpoint_file)
+
+# getting image prefix and test dataset from config file
+img_prefix = model.cfg["data"]["test"]["img_prefix"]
+test_dataset = model.cfg["data"]["test"]["ann_file"]
+with jsonlines.open(test_dataset, 'r') as fp:
+    test_file = list(fp)
+
+# generate prediction of html and save result to savepath
+pred_dict = dict()
+for sample in tqdm(test_file):
+    img_path = img_prefix + sample['filename']
+    pred_dict[sample['filename']] = inference_model(model, img_path)[0]
+with open(savepath, "w", encoding="utf-8") as writer:
+    json.dump(pred_dict, writer, ensure_ascii=False)
+
+# generate ground-truth of html from pubtabnet annotation of test dataset.
+gt_dict = dict()
+for data in test_file:
+    if data['filename'] in pred_dict.keys():
+        str_true = data['html']['structure']['tokens']
+        gt_dict[data['filename']] = {'html': format_html(data)}
+
+# evaluation using script from PubTabNet
+teds = TEDS(structure_only=False, n_jobs=16)
+scores = teds.batch_evaluate(pred_dict, gt_dict)
+print(np.array(list(scores.values())).mean())